requirements.txt
gradio>=4.0.0
pymupdf>=1.23.0
openai>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.2
markdown>=3.5.0